Fix db and session

+++ src/db/models/_role.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import ForeignKey
-from sqlalchemy.orm import Mapped, mapped_column, relationship
+from sqlalchemy.orm import Mapped, mapped_column
 
 from db.base import Base, int64, int64_pk, str_256
 
@@ -13,4 +12,4 @@ class Role(Base):
 
     id: Mapped[int64_pk]
     name: Mapped[str_256] = mapped_column(unique=True)
-    weight: Mapped[int64] = mapped_column()
+    weight: Mapped[int] = mapped_column()
